+++ src/hooks.rs
@@ -38,6 +38,8 @@ use crate::{
     refresh::{RefreshRegistry, TaskType},
 };
 
+use crate::param_utils::IntoProviderParam;
+
 pub use crate::provider_state::ProviderState;
 
 /// A unified trait for defining providers - async operations that return data
@@ -312,76 +314,6 @@ pub fn use_clear_provider_cache() -> impl Fn() + Clone {
     move || {
         cache.clear();
         refresh_registry.clear_all();
-    }
-}
-
-/// Trait for normalizing different parameter formats to work with providers
-///
-/// This trait allows the `use_provider` hook to accept parameters in different formats:
-/// - `()` for no parameters
-/// - `(param,)` for single parameter in tuple
-/// - Common primitive types directly
-///
-/// This eliminates the need for multiple `UseProvider` implementations.
-pub trait IntoProviderParam {
-    /// The target parameter type after conversion
-    type Param: Clone + PartialEq + Hash + Debug + Send + Sync + 'static;
-
-    /// Convert the input into the parameter format expected by the provider
-    fn into_param(self) -> Self::Param;
-}
-
-// Implementation for no parameters: () -> ()
-impl IntoProviderParam for () {
-    type Param = ();
-
-    fn into_param(self) -> Self::Param {
-        ()
-    }
-}
-
-// Implementation for tuple parameters: (Param,) -> Param
-impl<T> IntoProviderParam for (T,)
-where
-    T: Clone + PartialEq + Hash + Debug + Send + Sync + 'static,
-{
-    type Param = T;
-
-    fn into_param(self) -> Self::Param {
-        self.0
-    }
-}
-
-// Common direct parameter implementations to avoid conflicts
-impl IntoProviderParam for u32 {
-    type Param = u32;
-
-    fn into_param(self) -> Self::Param {
-        self
-    }
-}
-
-impl IntoProviderParam for i32 {
-    type Param = i32;
-
-    fn into_param(self) -> Self::Param {
-        self
-    }
-}
-
-impl IntoProviderParam for String {
-    type Param = String;
-
-    fn into_param(self) -> Self::Param {
-        self
-    }
-}
-
-impl IntoProviderParam for &str {
-    type Param = String;
-
-    fn into_param(self) -> Self::Param {
-        self.to_string()
     }
 }
