+++ src/lib.rs
@@ -7,6 +7,7 @@ pub mod global;
 pub mod hooks;
 pub mod injection;
 pub mod mutation;
+mod param_utils;
 pub mod platform;
 mod provider_state;
 pub mod refresh;
